requirements.txt
gunicorn
flask
flask-cors
dotenv
openpyxl
opencv-python
scikit-learn
scikit-image
matplotlib
pandas
openslide-python
openslide-bin
tqdm
slideio
# tensorflow
pillow
pymysql
Flask-Migrate

# FOR GPU CONFIG..
numpy==1.26.4
# torch==2.3.0+cu121
# torchvision==0.18.0+cu121
# torchaudio==2.3.0+cu121
# --extra-index-url https://download.pytorch.org/whl/cu121